+++ game/entities/pickups.h
@@ -11,7 +11,7 @@ class Pickups
     static ObjectPool<Pickups, 8> s_temporary;
     static ObjectPool<Pickups, 6> s_special;
 
-    static std::vector<uint16_t> s_acquired_specials; // store map indices of specials that have already been acquired
+    static std::vector<int> s_acquired_specials; // store map indices of specials that have already been acquired
 
     Vec2i position{0,0};
     SpriteWrapper m_sprite;
